move sort to table processing
